Promote helper functions to modules
