+++ lib/get_persister_by_path.js
@@ -1,0 +1,5 @@
+var getPersisterByPath = function (dataStore, path) {
+  return dataStore[path].persister;
+};
+
+module.exports = getPersisterByPath;
