+++ lib/get_state_by_path.js
@@ -1,0 +1,7 @@
+var ObjectPath = require('object-path');
+
+var getStateByPath = function (state, path) {
+  return ObjectPath.get(state, path);
+};
+
+module.exports = getStateByPath;
