+++ lib/put_cursor_on_chan.js
@@ -1,0 +1,9 @@
+var csp = require("js-csp");
+var go = csp.go;
+var put = csp.put;
+
+var putCursorOnChan = function (ch, cur) {
+  go(function* () { yield put(ch, cur); });
+};
+
+module.exports = putCursorOnChan;
